added hybrid memory option (a.k.a. manual memory timer)
Hybrid memory is a combination of automatic and manual, depending
on how long the light was off.  If it was a short time, automatic
mem is used.  For a longer time, it resets to the manual memory level.

Also tweaked some of the larger build targets again to make sure
they fit in ROM.