@hank-*-boost: reduced ripple on low modes by raising MCU clock speed to half at levels 2+ instead of the previous value of 1/4th speed
I tried full speed too, which makes ripple much smaller and faster...
but it also causes a big jump in brightness between levels 1 and 2.
My lux meter shows ~350 at 1/150 or ~500 at 2/150,
but at half speed it's ~650 at 2/150,
and at full speed it's ~1100 at 2/150.
So I went for a happy medium to balance ripple, brightness, and runtime.